changed fields of "likes" and "shares" from int to long

+++ src/main/java/eu/socialsensor/framework/client/search/solr/SolrItem.java
@@ -314,7 +314,7 @@ public class SolrItem {
     private int negativeVotes;
     
     @Field(value = "retweetsCount")
-    private Integer retweetsCount = 0;
+    private Long retweetsCount = 0L;
     
     
     public int getPositiveVotes() {
